webchat first commit
This website was my project that I did in college for Higher Diploma in
Science in Web Development @ Griffith College Cork. Since finishing the
course, I decided to continue with the website and continue.

I made a copy of the files and the first changes I made are to do with
style color and added in javascript code to welcome users to the site